# Fix XXH3 streaming performance