@Cyan4973 2021-11-27
Strange,
why would this change fail sanity tests 
specifically on `PPC64LE` + `Power8 Vector` ?