@easyaspi314 2021-11-28
> edit 4 : so, apparently, the vsx code path is really sensible to accumulators' alignment, so that should fix it.
> Not sure why older commits have started to fail too, but well, since #597 fixes that too ...

GCC for the undermaintained architectures is an untamed beast ¯\\\_(ツ)\_/¯